More IdPV3 compat rewrites - LDAP connections
The existing code was parsing the IdP attribute resolver configuration file (outside the IdP/spring).

This did not work well with the new IdPV3 style of using properties in the attribute resolver config file - they would not get substituted.

Also, the whole approach of parsing the attribute-resolver.xml file outside the
IdP and openning an LDAP connection independently was error prone (see e.g. #2 and #3).

While these have been fixed, ditch the code anyway and switch to directly using
the same connection profile that the LDAPDataConnector has already parsed.

The changes introduced in this commit are:
* SharedTokenDataConnector.storeSharedToken now stores the value directly by using LDAPDataConnector connectionFactory and searchBuilder.
* LdapUtil helper class is no longer needed and has been removed.
* conf/sharedtoken.properties is no longer needed (all info now accessed from LDAPDataConnector) and has been removed.
* Properties idpHome and subtreeSearch on SharedDataConnector are no longer needed and have been removed.
* The corresponding idpHome and subtreeSearch attributes in schema/arcs-shibext-dc.xsd have also been removed.

* POM.xml now has a dependency on idp-attribute-resolver-impl (3.0.0) - unfortunately necessary for the above approach to work.
 * Still better then error prone XML parsing + duplicating LDAP connection code.