Release 1.6.0 for IdP 3.0.0
See previous commits for details on what changed.

The code has been largely rewritten to match the IdPV3 API.

Unfortunately, the code also depends on IdPV3 Resolver Implementation
(LDAPDataConnector) - but the upside is that we now reuse the LDAP connection
the resolver already makes and the code parsing the attribute resolver
configuration file independently is gone.