Fix attribute dependency handling (+misc fixes)
Partially addresses #4

The SharedTokenDataConnector depends on the source attribute(s) (where it gets
the user's local ID from) - but this dependency is undeclared - and as of
1.6.0, the connector insists on having exactly one dependency, the LDAP data
connector.

As of Shibboleth IdP V3, the LDAP no longer automatically creates all
attributes it needs and (depending on the order of evaluation), undeclared
dependencies may be missed.

This results into sharedToken creation failing (because local IDs are not
found), or possibly it could result into the value being re-created (if the
existing sharedToken value in LDAP was missed).

Fix this with the following changes:

* Allow multiple dependencies - and recommend to have the source attribute as a dependency
 * update documentation
* specify the ID of the connector (for storing in ldap) as an extra property: ldapConnectorId
 * add to schema as a bean attribute
 * add to bean parser and pass as a bean property
* remove constraint on exactly one dependency
 * add meaningful checks instead:
  * if storeLdap is true, check ldapConnectorId is set AND is found among dependencies
  * issue a warning if sourceAttributeId is not found among dependencies
  * if storeDatabase is true, check stStore is set.
* check connector lookup succeeded if storeSharedToken
 * rename storeSharedToken to storeSharedTokenInLdap
* sourceAttributeId may be a comma separated list - honour this in validation
 * and in log messages elsewhere in the code
* move validatation code from validate() to doInitialize()
 * and remove abandoned validate() method (relict from IdPV2 API)

TODO: to fully address #4, we need also a solution for dependencies on the
auEduPersonSharedToken attribute fetched from LDAP.