SharedTokenDC: get LDAP attibute from LDAP DC
Get the existing sharedToken value from the LDAP connector directly (from the list of ResolvedAttributes on the DC directly).

We cannot rely on just getting STORED_ATTRIBUTE_NAME from resolverWorkContext.getResolvedIdPAttributeDefinitions()
- because there won't be an attribute definition of the same name to resolve from LDAP.

(Such a definition would have an ID clashing with the attribute definition done using this connector.)

Yes, we could import the attribute from LDAP explicitly under a different name
And then pass that name to this connector as an additional parameter.
But for now, let's get STORED_ATTRIBUTE_NAME as a ResolvedAttribute from the sharedTokenDC directly.

Also, add a check in doInitialize to make sure ldapConnectorId is set and the
connector provided as a dependency even when not storing into LDAP.

Also, improve how/when the checks are done and warnings issued - and also log a warning when both storeLdap and storeDatabase are set.

Fully fixes #4.