Logging: avoid logging raw binary data
Workaround for AD setups where local ID would be based on objectGUID - which is
binary but the LDAP configuration not always renders it as such - and would
then be putting binary data into the log file.

If the local ID is just alphanumeric, log the local ID unchanged.

Otherwise, log the local ID base 64 encoded.