fix: look up source attribute values also against resolvedDataConnectors
Previously, the code was looking only at resolved attribute definitions - but
if the attribute is not listed as an InputAttributeDefinition (only via an
InputDataConnector), the attribute definition might not be resolved yet.

Check and only use connectors explicitly listed as Inputs.

But do not apply this filtering to attribute definitions - as listing them has not been enforced
(only a warning emitted from doInitialize), so this could break existing code.

Make this check mandatory in IdP 4.x.